separated examples into own files
