+++ examples/config.ts
@@ -1,0 +1,2 @@
+export const colorItems = ['orange', 'purple', 'blue', 'red'];
+export const duration = 500;
